+++ client/src/redux/actions/actions.js
@@ -1,49 +1,18 @@
 import axios from "axios";
+import { GET_ARTICLE_TAG } from "./actionsArticles";
 
 const { REACT_APP_URL_API } = process.env;
 
 export const GET_ALL_CATEGORIES = "GET ALL CATEGORIES";
-export const GET_ALL_ARTICLE = "GET ALL ARTICLE";
-export const GET_ARTICLE_DETAIL = "GET ARTICLE DETAIL";
 export const GET_USERS = "GET USERS";
-export const GET_ARTICLE_TAG = "GET_ARTICLE_TAG";
 export const SET_USER_ID = "SET_USER_ID";
 export const SET_USER_ROLES = "SET_USER_ROLES";
-export const ORDER_ARTICLES = "ORDER_ARTICLES";
 export const GET_ALL_CAT_SUB = "GET_ALL_CAT_SUB";
 export const GET_CATEGORY = "GET_CATEGORY";
 export const GET_SUB_CATEGORY = "GET_SUB_CATEGORY";
 export const GET_USERS_BY_ROLES = "GET_USERS_BY_ROLES";
 export const GET_ADMINS = "GET_ADMINS";
 
-export const getAllArticle = () => {
-  return async (dispatch) => {
-    try {
-      const response = await axios(`${REACT_APP_URL_API}/article`);
-      dispatch({
-        type: GET_ALL_ARTICLE,
-        payload: response.data,
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-};
-
-export const getArticleDetail = (id) => (dispatch) => {
-  axios
-    .get(`${REACT_APP_URL_API}/article/${id}`)
-    .then((respuesta) => {
-      dispatch({ type: GET_ARTICLE_DETAIL, payload: respuesta.data });
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-};
-
-export const clearDetail = () => {
-  return { type: GET_ARTICLE_DETAIL, payload: undefined };
-};
 
 export const getAllUsers = () => {
   return async (dispatch) => {
@@ -292,32 +261,6 @@ export const deletePost = (art_id, token) => async (dispatch) => {
   }
 };
 
-export const orderArticles = (orderBy, order) => {
-  return async (dispatch) => {
-    try {
-      const response = await axios.get(
-        `${REACT_APP_URL_API}/article?orderBy=${orderBy}&order=${order}`
-      );
-      dispatch({
-        type: ORDER_ARTICLES,
-        payload: response.data,
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-};
-
-export function getArticleTag(tag) {
-  return function (dispatch) {
-    return axios
-      .get(`${REACT_APP_URL_API}/search?search=${tag}`)
-      .then((response) => response.data)
-      .then((json) => {
-        dispatch({ type: "GET_ARTICLE_TAG", payload: json });
-      });
-  };
-}
 
 export const getAllCatSub = () => {
   return async (dispatch) => {
