+++ client/src/redux/reducers/index.js
@@ -4,11 +4,13 @@ import  rootReducer  from './reducer';
 import institutionsReducer from './institutionsReducer';
 import usersReducer from './usersReducer';
 import metadataReducer from './metadataReducer';
+import articlesReducer from './articlesReducer'
 
 // combina todos los reducer
 export default combineReducers({
     metadataReducer,
     rootReducer,
     institutionsReducer,
-    usersReducer
+    usersReducer,
+    articlesReducer
 })
