+++ client/src/redux/reducers/reducer.js
@@ -1,6 +1,4 @@
 import {
-  GET_ALL_ARTICLE,
-  GET_ARTICLE_DETAIL,
   GET_USERS,
   SET_USER_ID,
   SET_USER_ROLES,
@@ -13,12 +11,10 @@ import {
 } from "../actions/actions";
 
 const initialState = {
-  articles: [],
-  articlesDetail: undefined,
   users: [],
   user_id: undefined,
   user_roles: [],
-  orderedArticles: [],
+  
   cat_sub: {},
   category: {},
   subCategory:{},
@@ -27,16 +23,6 @@ const initialState = {
 
 function rootReducer(state = initialState, action) {
   switch (action.type) {
-    case GET_ALL_ARTICLE:
-      return {
-        ...state,
-        articles: action.payload,
-      };
-    case GET_ARTICLE_DETAIL:
-      return {
-        ...state,
-        articlesDetail: action.payload,
-      };
     case GET_USERS:
       return {
         ...state,
@@ -52,16 +38,8 @@ function rootReducer(state = initialState, action) {
         ...state,
         user_roles: action.payload,
       };
-    case ORDER_ARTICLES:
-      return {
-        ...state,
-        orderedArticles: action.payload,
-      };
-    case GET_ARTICLE_TAG:
-      return {
-        ...state,
-        orderedArticles: action.payload,
-      };
+    
+    
     case GET_ALL_CAT_SUB:
       return {
         ...state,
